Add `useFooterActionsMeasurements` hook
