+++ ts/screens/wallet/payment/TransactionSummaryScreen.tsx
@@ -66,6 +66,7 @@ import {
   zendeskPaymentOrgFiscalCode,
   zendeskPaymentStartOrigin
 } from "../../../utils/supportAssistance";
+import { useFooterActionsMeasurements } from "../../../hooks/useFooterActionsMeasurements";
 import { TransactionSummary } from "./components/TransactionSummary";
 import { TransactionSummaryErrorDetails } from "./components/TransactionSummaryErrorDetails";
 import { TransactionSummaryStatus } from "./components/TransactionSummaryStatus";
@@ -155,17 +156,8 @@ const TransactionSummaryScreen = (): React.ReactElement => {
   } = useIOSelector(state => state.wallet.payment);
 
   /* Get `FooterActions` measurements */
-  const [footerActionsMeasurements, setfooterActionsMeasurements] =
-    useState<FooterActionsMeasurements>({
-      actionBlockHeight: 0,
-      safeBottomAreaHeight: 0
-    });
-
-  const handleFooterActionsMeasurements = (
-    values: FooterActionsMeasurements
-  ) => {
-    setfooterActionsMeasurements(values);
-  };
+  const { footerActionsMeasurements, handleFooterActionsMeasurements } =
+    useFooterActionsMeasurements();
 
   const error: TransactionSummaryError = pot.isError(paymentVerification)
     ? O.some(paymentVerification.error)
